BennettAlert: Work Save Day

+++ app/src/main/java/com/gautam/bennettalert/NewQuery Database/Alert.kt
@@ -5,7 +5,7 @@ import androidx.room.*
 @Entity
 data class Alert(
     @PrimaryKey(autoGenerate = true)
-    val queryId: Int,
+    val queryId: Long,
     val queryDetails: String,
     val queryAddress: String,
     val queryPriorityHigh: Boolean
